pyproject.toml
[project]
name = "tg-inviter-pro"
version = "0.1.0"
description = "Lightweight Telegram Inviter with Multi-account support and Web UI"
requires-python = ">=3.13"
dependencies = [
    "kurigram>=2.2.18",
    "tortoise-orm[sqlite]>=0.24.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "dishka>=1.2.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.12",
    "aiocsv>=1.3.0",
    "aiofiles>=24.1.0",
    "pydantic-settings>=2.2.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.0.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
